Move entitlement processor to a separate file so it can be made more complex without adding complexity to the entitlement subsystem header
